More hacks to allow compilation with newer compilers
A load of horrible hacks that allow compilation with 64 bit compilers.
It's all a bit embarassing.

+++ client.cpp
@@ -3144,7 +3144,7 @@ void Client::commitProtocolToDB(Protocol* protocol){
   QString cleanDescription = protocol->description.c_str();
   cleanText2(cleanName, "\\|");
   cleanText2(cleanDescription, "\\|");
-  *t << "<commitProtocolToDB>" << (int)protocol << "|"         // so that we can have some sort of identifier.. for acknowledgment purposes.. 
+  *t << "<commitProtocolToDB>" << (int)(long)protocol << "|"         // so that we can have some sort of identifier.. for acknowledgment purposes.. 
      << protocol->parentId << "|"
      << cleanName << "|" << cleanDescription << "|"
      << protocol->steps.size() << "|";
